# src/VMTO.API/Endpoints/ConnectionEndpoints.cs
using VMTO.API.Auth;
using VMTO.API.Extensions;
using VMTO.Application.Commands;
using VMTO.Application.Commands.Connections;
using VMTO.Application.DTOs;
using VMTO.Application.Ports.Repositories;
using VMTO.Application.Ports.Services;
using VMTO.Application.Queries;
using VMTO.Application.Queries.Connections;
using VMTO.Domain.Aggregates.Connection;

namespace VMTO.API.Endpoints;

public static class ConnectionEndpoints
{
    public static void MapConnectionEndpoints(this IEndpointRouteBuilder app)
    {
        var group = app.MapGroup("/api/connections").WithTags("Connections").RequireAuthorization();

        // 讀取操作 — 任何已認證使用者皆可存取
        group.MapGet("/", ListConnections);
        group.MapGet("/{id:guid}", GetConnection);
        group.MapGet("/{id:guid}/vms", ListVms);
        group.MapGet("/{id:guid}/vms/{vmId}", GetHyperVVmDetails);

        // 寫入操作 — 僅限 Admin 或 Operator，套用寫入速率限制
        group.MapPost("/", CreateConnection).RequireAuthorization(policy =>
            policy.RequireRole(Roles.Admin, Roles.Operator)).RequireRateLimiting("write");
        group.MapPost("/{id:guid}/validate", ValidateConnection).RequireAuthorization(policy =>
            policy.RequireRole(Roles.Admin, Roles.Operator)).RequireRateLimiting("write");
        group.MapPost("/{id:guid}/vms/{vmId}/preflight", RunPreFlightCheck).RequireAuthorization(policy =>
            policy.RequireRole(Roles.Admin, Roles.Operator)).RequireRateLimiting("write");
        group.MapDelete("/{id:guid}", DeleteConnection).RequireAuthorization(policy =>
            policy.RequireRole(Roles.Admin, Roles.Operator)).RequireRateLimiting("write");
    }

    private static async Task<IResult> ListConnections(
        IConnectionRepository repo,
        int page = 1,
        int pageSize = 20,
        CancellationToken ct = default)
    {
        pageSize = Math.Min(pageSize, 100);
        var connections = await repo.ListAsync(page, pageSize, ct);
        var total = await repo.CountAsync(ct);
        return Results.Ok(new { items = connections.Select(MapToDto), total, page, pageSize });
    }

    private static async Task<IResult> GetConnection(Guid id, IConnectionRepository repo, CancellationToken ct)
    {
        var connection = await repo.GetByIdAsync(id, ct);
        if (connection is null) return Results.NotFound();
        return Results.Ok(MapToDto(connection));
    }

    private static async Task<IResult> CreateConnection(
        CreateConnectionRequest request,
        IConnectionRepository repo,
        IEncryptionService encryption,
        CancellationToken ct)
    {
        var encryptedSecret = encryption.Encrypt(request.Secret);
        var connection = new Connection(request.Name, request.Type, request.Endpoint, encryptedSecret, request.MetadataJson);
        await repo.AddAsync(connection, ct);
        return Results.Created($"/api/connections/{connection.Id}", MapToDto(connection));
    }

    private static async Task<IResult> ValidateConnection(
        Guid id,
        ICommandHandler<ValidateConnectionCommand> handler,
        HttpContext context,
        CancellationToken ct)
    {
        var result = await handler.HandleAsync(new ValidateConnectionCommand(id), ct);
        return result.ToHttpResult(context);
    }

    private static async Task<IResult> ListVms(
        Guid id,
        IQueryHandler<ListVmsQuery, IReadOnlyList<VmInfoDto>> handler,
        HttpContext context,
        CancellationToken ct)
    {
        var result = await handler.HandleAsync(new ListVmsQuery(id), ct);
        return result.ToHttpResult(context);
    }

    private static async Task<IResult> GetHyperVVmDetails(
        Guid id,
        string vmId,
        IQueryHandler<GetHyperVVmDetailsQuery, HyperVVmDetailsDto> handler,
        HttpContext context,
        CancellationToken ct)
    {
        var result = await handler.HandleAsync(new GetHyperVVmDetailsQuery(id, vmId), ct);
        return result.ToHttpResult(context);
    }

    private static async Task<IResult> RunPreFlightCheck(
        Guid id,
        string vmId,
        ICommandHandler<RunPreFlightCheckCommand, PreFlightCheckResultDto> handler,
        HttpContext context,
        CancellationToken ct)
    {
        var result = await handler.HandleAsync(new RunPreFlightCheckCommand(id, vmId), ct);
        return result.ToHttpResult(context);
    }

    private static async Task<IResult> DeleteConnection(Guid id, IConnectionRepository repo, IJobRepository jobRepo, CancellationToken ct)
    {
        var connection = await repo.GetByIdAsync(id, ct);
        if (connection is null) return Results.NotFound();
        if (await jobRepo.HasActiveJobsForConnectionAsync(id, ct))
            return Results.Conflict(new { error = "Cannot delete connection: there are active jobs using this connection." });
        await repo.DeleteAsync(id, ct);
        return Results.NoContent();
    }

    private static ConnectionDto MapToDto(Connection c) =>
        new(c.Id, c.Name, c.Type, c.Endpoint, c.ValidatedAt, c.CreatedAt);
}

public sealed record CreateConnectionRequest(
    string Name,
    PlatformKind Type,
    string Endpoint,
    string Secret,
    string? MetadataJson = null);