# src/VMTO.API/Endpoints/JobEndpoints.cs
using VMTO.API.Auth;
using VMTO.Application.DTOs;
using VMTO.Application.Ports.Repositories;
using VMTO.Domain.Aggregates.MigrationJob;
using VMTO.Domain.Enums;

namespace VMTO.API.Endpoints;

public static class JobEndpoints
{
    public static void MapJobEndpoints(this IEndpointRouteBuilder app)
    {
        var group = app.MapGroup("/api/jobs").WithTags("Jobs").RequireAuthorization();

        // 讀取操作 — 任何已認證使用者皆可存取
        group.MapGet("/", ListJobs);
        group.MapGet("/{id:guid}", GetJob);
        group.MapGet("/{id:guid}/progress", GetJobProgress);

        // 寫入操作 — 僅限 Admin 或 Operator，套用寫入速率限制
        group.MapPost("/", CreateJob).RequireAuthorization(policy =>
            policy.RequireRole(Roles.Admin, Roles.Operator)).RequireRateLimiting("write");
        group.MapPost("/{id:guid}/cancel", CancelJob).RequireAuthorization(policy =>
            policy.RequireRole(Roles.Admin, Roles.Operator)).RequireRateLimiting("write");
        group.MapPost("/{id:guid}/pause", PauseJob).RequireAuthorization(policy =>
            policy.RequireRole(Roles.Admin, Roles.Operator)).RequireRateLimiting("write");
        group.MapPost("/{id:guid}/resume", ResumeJob).RequireAuthorization(policy =>
            policy.RequireRole(Roles.Admin, Roles.Operator)).RequireRateLimiting("write");
        group.MapPost("/{id:guid}/retry", RetryFailedSteps).RequireAuthorization(policy =>
            policy.RequireRole(Roles.Admin, Roles.Operator)).RequireRateLimiting("write");
    }

    private static async Task<IResult> ListJobs(
        IJobRepository repo,
        int page = 1,
        int pageSize = 20,
        JobStatus? status = null,
        CancellationToken ct = default)
    {
        pageSize = Math.Min(pageSize, 100);
        var jobs = await repo.ListAsync(page, pageSize, status, ct);
        var total = await repo.CountAsync(status, ct);
        return Results.Ok(new { items = jobs.Select(MapToDto), total, page, pageSize });
    }

    private static async Task<IResult> GetJob(Guid id, IJobRepository repo, CancellationToken ct)
    {
        var job = await repo.GetByIdAsync(id, ct);
        if (job is null) return Results.NotFound();
        return Results.Ok(MapToDto(job));
    }

    private static async Task<IResult> CreateJob(
        CreateJobRequest request,
        VMTO.Application.Commands.ICommandHandler<VMTO.Application.Commands.Jobs.CreateJobCommand, Guid> handler,
        IJobRepository repo,
        MassTransit.IPublishEndpoint publish,
        HttpContext http,
        CancellationToken ct)
    {
        var command = new VMTO.Application.Commands.Jobs.CreateJobCommand(
            request.SourceConnectionId,
            request.TargetConnectionId,
            request.StorageTarget,
            request.Strategy,
            request.Options,
            request.VmId,
            request.DiskKeys);

        var result = await handler.HandleAsync(command, ct);
        if (!result.IsSuccess)
            return VMTO.API.Extensions.ResultExtensions.ToHttpResult(result, http);

        var job = await repo.GetByIdAsync(result.Value, ct);
        if (job is not null)
            await publish.Publish(VMTO.Application.Messages.JobStartedMessage.FromJob(job), ct);

        return Results.Created($"/api/jobs/{result.Value}", job is null ? new { id = result.Value } : MapToDto(job));
    }

    private static async Task<IResult> CancelJob(Guid id, IJobRepository repo, CancellationToken ct)
    {
        var job = await repo.GetByIdAsync(id, ct);
        if (job is null) return Results.NotFound();
        var result = job.RequestCancel();
        if (!result.IsSuccess) return Results.BadRequest(new { error = result.ErrorCode, message = result.ErrorMessage });
        await repo.UpdateAsync(job, ct);
        return Results.Ok(MapToDto(job));
    }

    private static async Task<IResult> PauseJob(Guid id, IJobRepository repo, CancellationToken ct)
    {
        var job = await repo.GetByIdAsync(id, ct);
        if (job is null) return Results.NotFound();
        var result = job.RequestPause();
        if (!result.IsSuccess) return Results.BadRequest(new { error = result.ErrorCode, message = result.ErrorMessage });
        await repo.UpdateAsync(job, ct);
        return Results.Ok(MapToDto(job));
    }

    private static async Task<IResult> ResumeJob(Guid id, IJobRepository repo, CancellationToken ct)
    {
        var job = await repo.GetByIdAsync(id, ct);
        if (job is null) return Results.NotFound();
        var result = job.RequestResume();
        if (!result.IsSuccess) return Results.BadRequest(new { error = result.ErrorCode, message = result.ErrorMessage });
        await repo.UpdateAsync(job, ct);
        return Results.Ok(MapToDto(job));
    }

    private static async Task<IResult> RetryFailedSteps(Guid id, IJobRepository repo, CancellationToken ct)
    {
        var job = await repo.GetByIdAsync(id, ct);
        if (job is null) return Results.NotFound();
        foreach (var step in job.Steps.Where(s => s.Status == StepStatus.Failed))
        {
            step.Retry();
        }
        await repo.UpdateAsync(job, ct);
        return Results.Ok(MapToDto(job));
    }

    private static async Task<IResult> GetJobProgress(Guid id, IJobRepository repo, CancellationToken ct)
    {
        var job = await repo.GetByIdAsync(id, ct);
        if (job is null) return Results.NotFound();
        var progress = new JobProgressDto(
            job.Id,
            job.Status,
            job.Progress,
            job.Steps.Select(MapStepToDto).ToList());
        return Results.Ok(progress);
    }

    private static JobDto MapToDto(MigrationJob job) =>
        new(job.Id,
            job.CorrelationId.Value,
            job.Strategy,
            job.Status,
            job.Progress,
            job.CreatedAt,
            job.UpdatedAt,
            job.Steps.Select(MapStepToDto).ToList());

    private static JobStepDto MapStepToDto(JobStep step) =>
        new(step.Id,
            step.Name,
            step.Order,
            step.Status,
            step.Progress,
            step.RetryCount,
            step.ErrorMessage);
}

public sealed record CreateJobRequest(
    Guid SourceConnectionId,
    Guid TargetConnectionId,
    Domain.ValueObjects.StorageTarget StorageTarget,
    MigrationStrategy Strategy,
    MigrationOptions Options,
    string VmId,
    IReadOnlyList<string>? DiskKeys = null);